Rollback inc_ref() calls made during partial reads.
Previously, if a relay received a turn like

    [[0 <A [#:[0 22]
       9
       9
       9
    ] 11>]]

across multiple packets - say, one per line depicted above - then it
would partially parse the input multiple times, each time calling
inc_ref() on the WireSymbol corresponding to #:[0 22]. This means that
the net effect, at the end of the read, would be (in this case) +5, when
it should have been +1 to the reference count.

This patch remembers (and performs) an obligation to roll back in case
of partial parse or other input error.

(There's still the "accidentally quadratic" problem of partial parses
to deal with.)